migrate test_autoscaling_with_streaming to test_autoscaling_policy.py
